reste to 9/18 hash via BB deployment, render with homepage update, Embed GA4 analytics and update book content
