+++ _quarto.yml
@@ -1,24 +1,31 @@
 project:
   type: book
+  output-dir: docs
 
 book:
-  title: "SERVIR's Applied Deep Learning Handbook"
+  title: "Applied Artificial Intelligence and Deep Learning Book"
   author:
     - name: Tim Mayer
       affiliations:
-        - id: nasa-servir
-          name: NASA - SERVIR
+        - id: nasa-earthtrise
+          name: NASA - EarthRISE
         - id: UofA-huntsville
           name: The University of Alabama in Huntsville
     - name: Biplov Bhandari
       affiliations:
         - id: woolpert
           name: Woolpert Digital Innovations
+    - name: David Saah
+      affiliations:
+        - id: USF
+          name: University of San Francisco
+        - id: spatial-informatics-group
+          name: Spatial Informatics Group
 
-  date: "11/15/2024"
+  date: "09/30/2024"
   cover-image: Images/Book_Cover_delta_6_25.png #images/Book_Cover.png
   # repo-url: https://servir.github.io/SERVIR-Applied-Deep-Learning-Book/ #https://mayert1.github.io/quarto_book_trial/
-  repo-url: https://github.com/SERVIR/SERVIR-Applied-Deep-Learning-Book
+  repo-url:  https://github.com/NASA-EarthRISE/EarthRISE-Applied-Artificial-Intelligence-and-Deep-Learning-Book
   repo-actions: [source, edit, issue]
   downloads: [pdf]
   chapters:
@@ -49,6 +56,7 @@ format:
     # output-fold: true
     # output-summary: "Show output"
     page-tools: ["colab", "github", "source", "edit", "issue"]
+    include-in-header: _includes/ga.html
   pdf:
     documentclass: scrreprt  # book
     page-tools: [] # Remove page tools for PDF
