+++ docs/03_Semantic_Segmentation/01__Crop_Mapping/notebooks/Rice_Mapping_Bhutan_2021_files/figure-html/cell-16-output-2.png
